+++ tests/Expect/Response/toBeForbidden.php
@@ -1,28 +1,36 @@
 <?php
 
-use function Pest\Laravel\get;
+use Illuminate\Http\Response;
 use PHPUnit\Framework\ExpectationFailedException;
 
 test('pass', function () {
-    $response = get('/secret');
+    $response = build_response(function (Response $response) {
+        $response->setStatusCode(403);
+    });
 
     expect($response)->toBeForbidden();
 });
 
 test('fails', function () {
-    $response = get('/ok');
+    $response = build_response(function (Response $response) {
+        $response->setStatusCode(200);
+    });
 
     expect($response)->toBeForbidden();
 })->throws(ExpectationFailedException::class, 'Expected response status code [403] but received 200');
 
 test('pass with negation', function () {
-    $response = get('/ok');
+    $response = build_response(function (Response $response) {
+        $response->setStatusCode(201);
+    });
 
     expect($response)->not->toBeForbidden();
 });
 
 test('fails with negation', function () {
-    $response = get('/secret');
+    $response = build_response(function (Response $response) {
+        $response->setStatusCode(403);
+    });
 
     expect($response)->not->toBeForbidden();
 })->throws(ExpectationFailedException::class, "Expecting Illuminate\Testing\TestResponse Object (...) not to be forbidden");