+++ docs/content/en/changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 **Added**
 
 - `toBeSameDayAs()` expectation
+- `toBeSameHourAs()` expectation
 - `toBeSameYearAs()` expectation
 
 ### [v1.4.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.3.0...v1.4.0) - 2021-10-21