+++ docs/content/en/changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `toBeSameDayAs()` expectation
 - `toBeSameHourAs()` expectation
+- `toBeSameMonthAs()` expectation
 - `toBeSameYearAs()` expectation
 
 ### [v1.4.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.3.0...v1.4.0) - 2021-10-21