+++ docs/content/en/changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 **Added**
 
+- `toBeAfter()` expectation
 - `toBeSameDayAs()` expectation
 - `toBeSameHourAs()` expectation
 - `toBeSameMonthAs()` expectation